backend/requirements.txt
fastapi==0.95.1
uvicorn==0.22.0
pydantic==1.10.7
python-dotenv==1.0.0
supabase==1.0.3
openai==1.3.0
python-multipart==0.0.6
